requirements.txt
# Core dependencies
nvidia-morpheus-core==25.06
tritonclient[all]==2.49.0
cudf-cu12==24.10.*
cuml-cu12==24.10.*
cugraph-cu12==24.10.*

# Data processing
pandas==2.2.0
numpy==1.26.4
pyarrow==15.0.0

# Kafka
confluent-kafka==2.3.0

# Monitoring
prometheus-client==0.19.0

# CLI
click==8.1.7
pyyaml==6.0.1
